fix kv cache dtype

+++ models/tt_transformers/tt/attention.py
@@ -718,7 +718,7 @@ class Attention(LightweightModule):
             keys_BKSD, values_BKSD = kv_cache[0], kv_cache[1]
         else:
             keys_BKSD, values_BKSD = self.layer_past[0], self.layer_past[1]
-        k_heads_1KSD_8b = ttnn.typecast(k_heads_1KSD, dtype=self.keys_BKSD.dtype)
+        k_heads_1KSD_8b = ttnn.typecast(k_heads_1KSD, dtype=keys_BKSD.dtype)
         ttnn.deallocate(k_heads_1KSD)
 
         # sharding k_fill to deal with update_cache memory limitation
